Parse labels separately from Flex and Bison
Label-detection is the only part of the syntax that is
whitespace-sensitive, but making the whole grammar
whitespace-sensitive is a mess, so I want to avoid that.

Parsing labels separately also helps for parsing macros, which must
be done entirely without Flex and Bison's involvement, as macros are
built around direct string manipulation rather than tokens and rules.