Fix shift/rotate instructions being too large
The static versions of these instruction weren't inlining their
immediate value properly.

This fixes a branch in Sonic 1's Kosinski decompressor being out of
range.

Unfortunately, it looks like the next (and final?) hurdle to
assembling the Sonic 1 disassembly is to support its multitude of
equates that use then-undefined symbols. I swear this is going to
require a multi-pass solution...