Make macro argument substitution stricter
This behavior is closer to that of AS, and is required by the Sonic 1
disassembly for its 'out_of_range' macro.

The idea is that a macro argument should not be inserted if the
placeholder that it is replacing appears to be a substring of a
larger identifier. This prevents an argument called 'pos' from
matching the first three letters of the word 'position'.