Add support for outputting a symbol (.sym) file
asm68k can generate one of these if you pass it a path between the
paths of the output binary file and the listing file.

https://sonicresearch.org/community/index.php?threads/advanced-error-handler-and-debugger.4442/

Vladikcomper's Advanced Error Handler uses this file to produce its
internal symbol list, so reproducing this feature is important for
supporting projects that rely on this error handler.