Allow constants to be redefined sometimes
asm68k allows constants to be redefined so long as they are redefined
to the value that they already are. This might sound pointless, but I
think the idea is to accomodate for redundant definitions that may be
produced by things like macros.

SuperEgg's Sonic 2 Nick Arcade disassembly relies on this.