Improve support for string/character literals
It turns out that 'string' and "string" are the same thing in asm68k.

String literals and character literals are the same thing - they can
just be interpreted differently depending on what is using them.

SuperEgg's Sonic 2 Nick Arcade disassembly wraps string literals in
"'" symbols, so this behaviour needs to implemented to support it.