Use a flexible array member for the dictionary
Unlike a zero-sized array (which is a GCC extension), or a one-sized
array (which is apparently undefined behaviour), this should be
perfectly standards-compliant, while still supporting C89.

This saves a few bytes per dictionary entry, as there no longer needs
to be a redundant identifier pointer, which just pointed to the end
of the struct.