Fix detection of split lines.
If a line ends with '&', then it is continued on the next line.

However, if the '&' exists as part of a comment, then S.N. 68k
(asm68k) will ignore it. This is relied on by the Sonic Spinball
disassembly. It also turns out that 68 Katy's BASIC interpreter
relies on this too, causing the resulting binary to be incorrect.